tr1/game-flow: split into modules
