+++ src/tr1/game/game_flow/common.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <libtrx/game/game_flow/common.h>
+
+void GF_Shutdown(void);
