+++ src/tr1/game/game_flow/reader.h
@@ -1,0 +1,3 @@
+#pragma once
+
+bool GF_Load(const char *file_name);
