+++ src/tr1/game/game_flow/vars.c
@@ -1,0 +1,3 @@
+#include "game/game_flow/vars.h"
+
+GAME_FLOW g_GameFlow = {};
