+++ src/tr1/global/types.h
@@ -173,36 +173,6 @@ typedef enum {
 } SLIDING_PILLAR_STATE;
 
 typedef enum {
-    GFS_START_GAME,
-    GFS_LOOP_GAME,
-    GFS_STOP_GAME,
-    GFS_START_CINE,
-    GFS_LOOP_CINE,
-    GFS_PLAY_FMV,
-    GFS_LEVEL_STATS,
-    GFS_TOTAL_STATS,
-    GFS_LOADING_SCREEN,
-    GFS_DISPLAY_PICTURE,
-    GFS_EXIT_TO_TITLE,
-    GFS_EXIT_TO_LEVEL,
-    GFS_EXIT_TO_CINE,
-    GFS_SET_CAM_X,
-    GFS_SET_CAM_Y,
-    GFS_SET_CAM_Z,
-    GFS_SET_CAM_ANGLE,
-    GFS_FLIP_MAP,
-    GFS_REMOVE_GUNS,
-    GFS_REMOVE_SCIONS,
-    GFS_GIVE_ITEM,
-    GFS_PLAY_SYNCED_AUDIO,
-    GFS_MESH_SWAP,
-    GFS_REMOVE_AMMO,
-    GFS_REMOVE_MEDIPACKS,
-    GFS_SETUP_BACON_LARA,
-    GFS_LEGACY,
-} GAME_FLOW_SEQUENCE_EVENT_TYPE;
-
-typedef enum {
     BT_LARA_HEALTH = 0,
     BT_LARA_MAX_AIR = 1,
     BT_ENEMY_HEALTH = 2,
