+++ src/tr1/meson.build
@@ -113,7 +113,10 @@ sources = [
   'game/game/game.c',
   'game/game/game_draw.c',
   'game/game_string.c',
-  'game/game_flow.c',
+  'game/game_flow/reader.c',
+  'game/game_flow/common.c',
+  'game/game_flow/sequencer.c',
+  'game/game_flow/vars.c',
   'game/gun/gun.c',
   'game/gun/gun_misc.c',
   'game/gun/gun_pistols.c',
