+++ src/tr2/game/game_flow/reader.c
@@ -68,13 +68,6 @@ static bool M_LoadDemos(JSON_OBJECT *obj, GAME_FLOW *gf);
 
 static void M_LoadTitleLevel(JSON_OBJECT *obj, GAME_FLOW *gf);
 
-static void M_FreeSequence(GAME_FLOW_SEQUENCE *sequence);
-static void M_FreeLevel(GAME_FLOW_LEVEL *level);
-static void M_FreeLevels(GAME_FLOW *gf);
-static void M_FreeDemos(GAME_FLOW *gf);
-static void M_FreeCutscenes(GAME_FLOW *gf);
-static void M_FreeFMVs(GAME_FLOW *gf);
-
 static M_SEQUENCE_EVENT_HANDLER m_SequenceEventHandlers[] = {
     // clang-format off
     // Events without arguments
@@ -525,73 +518,4 @@ end:
 
     Memory_FreePointer(&script_data);
     return result;
-}
-
-static void M_FreeSequence(GAME_FLOW_SEQUENCE *const sequence)
-{
-    Memory_Free(sequence->events);
-}
-
-static void M_FreeLevel(GAME_FLOW_LEVEL *const level)
-{
-    for (int32_t j = 0; j < level->injections.count; j++) {
-        Memory_FreePointer(&level->injections.data_paths[j]);
-    }
-    M_FreeSequence(&level->sequence);
-    Memory_FreePointer(&level->injections.data_paths);
-    Memory_FreePointer(&level->path);
-    Memory_FreePointer(&level->title);
-}
-
-static void M_FreeLevels(GAME_FLOW *const gf)
-{
-    for (int32_t i = 0; i < gf->level_count; i++) {
-        M_FreeLevel(&gf->levels[i]);
-    }
-    Memory_FreePointer(&gf->levels);
-    gf->level_count = 0;
-}
-
-static void M_FreeDemos(GAME_FLOW *const gf)
-{
-    Memory_FreePointer(&gf->demos);
-    gf->demo_count = 0;
-}
-
-static void M_FreeCutscenes(GAME_FLOW *const gf)
-{
-    for (int32_t i = 0; i < gf->cutscene_count; i++) {
-        Memory_FreePointer(&gf->cutscenes[i].path);
-    }
-    Memory_FreePointer(&gf->cutscenes);
-    gf->cutscene_count = 0;
-}
-
-static void M_FreeFMVs(GAME_FLOW *const gf)
-{
-    for (int32_t i = 0; i < gf->fmv_count; i++) {
-        Memory_FreePointer(&gf->fmvs[i].path);
-    }
-    Memory_FreePointer(&gf->fmvs);
-    gf->fmv_count = 0;
-}
-
-void GF_Shutdown(void)
-{
-    GAME_FLOW *const gf = &g_GameFlow;
-
-    for (int32_t i = 0; i < gf->injections.count; i++) {
-        Memory_FreePointer(&gf->injections.data_paths[i]);
-    }
-    Memory_FreePointer(&gf->injections.data_paths);
-
-    M_FreeLevels(gf);
-    M_FreeDemos(gf);
-    M_FreeCutscenes(gf);
-    M_FreeFMVs(gf);
-
-    if (gf->title_level != NULL) {
-        M_FreeLevel(gf->title_level);
-        Memory_FreePointer(&gf->title_level);
-    }
 }
